Test cleanup
We really only need two suites currently: A browser functional suite and
a node unit suite. All our tests can be added to sub-suties of those
top-level suites. We may eventually want to add a browser unit suite.
This change additionally adds descriptive comments at the top of each
suite to describe how to extend them and create new sub-suites, and
replaces a bunch of no-op functions that only existed as examples.
